BuildSession: consolidate compile pipeline behind one module
Pure consolidation, byte-identical outputs vs main (Stage A PR 1 of the
incremental-builds spec). lib/compiler/buildSession.ts now owns what was
scattered: the commands.ts module globals (compiledFiles/currentClosure),
ensureCompiledClosure, the directory union-closure branch, compileMany,
and precompile.ts cross-config machinery (grouping contract, configKey
derivation, single-slot assert). commands.ts keeps thin delegates
(compile/compileMany/resetCompilationCache) over one LAZY defaultSession;
precompile.ts shrinks to a test-runner adapter on a fresh session.

Notable:
- setClosure() makes the closure-replacement/dedupe-clear pairing
  structural instead of convention at three sites.
- The default session is created lazily: an eager top-level
  createBuildSession() defeated esbuild tree-shaking and dragged the
  whole codegen subtree into agency pack artifacts (caught by
  pack.test.ts, ~16k extra bundle lines).
- configKey is derived inside the session; the prebuilt-closure handoff
  is internal-only; readFile moved with the pipeline and is re-exported
  from commands.ts to avoid an import cycle.
