+++ packages/agency-lang/lib/cli/precompile.ts
@@ -1,5 +1,6 @@
 /**
- * Precompile pass for the Agency test runner.
+ * Precompile pass for the Agency test runner — a thin adapter over
+ * BuildSession.compileGroups.
  *
  * Historically each test CASE recompiled its `.agency` source (and the
  * source's whole import tree) via `executeNodeAsync` → `compile()` — with
@@ -8,40 +9,25 @@
  * unique source exactly once, up front; the runner then executes test cases
  * with `preferCompiled: true`, which reuses the sibling `.js`.
  *
- * Config grouping: compiled output is config-dependent (the generator bakes
- * config defaults into emitted code), and a test dir may carry a local
- * `agency.json` that `runTestFile` merges over the base config. Sources are
- * therefore grouped by merged config — one union-closure compile for all
- * base-config files, plus one per local-config dir.
- *
- * Cross-config invariant: a sibling `.js` is a single slot per module, so a
- * module reachable from two groups whose configs differ would be
- * last-writer-wins. That is asserted here and fails loudly (no test dir
- * does this today). A graceful fallback (per-case compiles for conflicting
- * files) was rejected: interleaved recompiles rewrite shared siblings
- * mid-run, which is exactly the race this pass removes.
+ * What lives HERE is only what is test-runner-specific: mapping `.test.json`
+ * files to sibling sources, honoring file-level skip/skipOnCI, and merging a
+ * dir-local `agency.json` over the base config. The config grouping
+ * contract, the cross-config single-slot assert, and the per-group compile
+ * loop live in lib/compiler/buildSession.ts.
  */
 import fs from "fs";
 import path from "path";
 import { AgencyConfig } from "../config.js";
-import { loadConfig, compileMany } from "./commands.js";
+import { loadConfig } from "./commands.js";
 import {
-  buildCompiledClosure,
-  CompileClosureError,
-  type CompiledClosure,
-} from "../compiler/compileClosure.js";
+  createBuildSession,
+  type CompileGroup,
+} from "../compiler/buildSession.js";
 
 const BASE_GROUP_LABEL = "<base config>";
 
-export type PrecompileGroup = {
-  /** Base-config marker or the local-`agency.json` dir, for error messages. */
-  label: string;
-  config: AgencyConfig;
-  /** Canonical serialization of `config`; groups conflict only when keys differ. */
-  configKey: string;
-  /** Absolute `.agency` entry paths. */
-  files: string[];
-};
+// Back-compat name for existing importers (precompile.test.ts).
+export type { CompileGroup as PrecompileGroup };
 
 // File-level skip mirror of runTestFile's check: a `skip: true` (or
 // `skipOnCI: true` under CI) .test.json never runs, so its source must not
@@ -59,10 +45,10 @@ function isFileLevelSkipped(testJsonFile: string): boolean {
 export function groupTestSources(
   baseConfig: AgencyConfig,
   testJsonFiles: string[],
-): PrecompileGroup[] {
+): CompileGroup[] {
   // Null-prototype: keyed by dir paths (see lib/optimize/registry.ts for the
   // house pattern on string-keyed registries).
-  const groups: Record<string, PrecompileGroup> = Object.create(null);
+  const groups: Record<string, CompileGroup> = Object.create(null);
   for (const testJsonFile of testJsonFiles) {
     const sourceFile = path
       .resolve(testJsonFile)
@@ -82,43 +68,11 @@ export function groupTestSources(
     const group = (groups[label] ??= {
       label,
       config,
-      // JSON.stringify is order-stable here: every config passes through
-      // AgencyConfigSchema.safeParse (loadConfigSafe), and zod rebuilds the
-      // object in SCHEMA shape order, not input order — so semantically
-      // identical configs always serialize identically. Guarded by the
-      // "configKey is key-order independent" test.
-      configKey: JSON.stringify(config),
       files: [],
     });
     if (!group.files.includes(sourceFile)) group.files.push(sourceFile);
   }
   return Object.values(groups);
-}
-
-export function findCrossConfigConflicts(
-  groups: { label: string; configKey: string; modules: string[] }[],
-): { module: string; labels: string[] }[] {
-  // Null-prototype: keyed by absolute module paths.
-  const touchedBy: Record<string, { configKey: string; label: string }[]> =
-    Object.create(null);
-  for (const group of groups) {
-    for (const module of group.modules) {
-      (touchedBy[module] ??= []).push({
-        configKey: group.configKey,
-        label: group.label,
-      });
-    }
-  }
-  const conflicts: { module: string; labels: string[] }[] = [];
-  for (const [module, touches] of Object.entries(touchedBy)) {
-    const distinctKeys = touches
-      .map((t) => t.configKey)
-      .filter((key, i, all) => all.indexOf(key) === i);
-    if (distinctKeys.length > 1) {
-      conflicts.push({ module, labels: touches.map((t) => t.label) });
-    }
-  }
-  return conflicts;
 }
 
 export function precompileTestSources(
@@ -127,37 +81,13 @@ export function precompileTestSources(
   options?: { quiet?: boolean },
 ): void {
   const groups = groupTestSources(baseConfig, testJsonFiles);
-  const withClosures: { group: PrecompileGroup; closure: CompiledClosure }[] =
-    groups.map((group) => ({
-      group,
-      closure: buildCompiledClosure(group.files, group.config),
-    }));
-
-  const conflicts = findCrossConfigConflicts(
-    withClosures.map(({ group, closure }) => ({
-      label: group.label,
-      configKey: group.configKey,
-      modules: Object.keys(closure.programs),
-    })),
-  );
-  if (conflicts.length > 0) {
-    const lines = conflicts.map(
-      (c) =>
-        `  ${c.module}\n    reachable from: ${c.labels.join(", ")}`,
-    );
-    throw new CompileClosureError(
-      "Test sources with differing configs share modules. A module's " +
-        "compiled .js is a single slot, so this would be last-writer-wins. " +
-        "Move the shared module or align the configs:\n" +
-        lines.join("\n"),
-    );
-  }
-
-  for (const { group, closure } of withClosures) {
-    compileMany(group.config, group.files, {
-      closure,
-      quiet: options?.quiet,
-      allowTestImports: true,
-    });
-  }
+  // Fresh session per precompile — a declared delta from the old shared
+  // module globals: the default session no longer inherits precompile's
+  // last-group closure/dedupe state. Safe because test cases run with
+  // `preferCompiled: true` and never re-enter compile(); the old inherited
+  // state was itself a latent staleness hazard.
+  createBuildSession().compileGroups(groups, {
+    quiet: options?.quiet,
+    allowTestImports: true,
+  });
 }
